# Add rich error messaging and re-entrant support for multi-segment json