--- a/src/System.Buffers.ReaderWriter/System/ThrowHelper.cs
+++ b/src/System.Buffers.ReaderWriter/System/ThrowHelper.cs
@@ -31,7 +31,7 @@ namespace System.Buffers
 
     internal enum ExceptionArgument
     {
-
+        count,
         length,
 
     }