@KrzysztofCwalina 2018-10-04
BufferReader will be used to write parsers, and parsers often need to backtrack. It seems like it would be super useful to support backtracking.