@KrzysztofCwalina 2018-10-04
The perf regression seems scary