@JeremyKuhne 2018-10-04
> It seems like it would be super useful to support backtracking.

We do in that you can copy the BufferReader. I wouldn't think you can do much better than that. @ahsonkhan, can you try to add a method that lets you set position and measure? Setting a long seems to be the only rational way to do it- we could optimize to see if we're in the current segment, otherwise we'd can fully reset and Advance.