@ahsonkhan 2018-10-04
> can you try to add a method that lets you set position and measure? 

I don't follow. Can you clarify what type of method you are thinking of? One that sets the current SequencePosition within BufferReader?