@JeremyKuhne 2018-10-04
> I don't follow. Can you clarify what type of method you are thinking of? One that sets the current SequencePosition within BufferReader?

``` C#
long consumed = reader.Consumed;
// ...
reader.SetPosition(consumed);
```

``` C#
public void SetPosition(long position)
{
    // Check for > 0
    // Is position in current segment?
    // Yes? Calculate CurrentSpanPosition and set & set Consumed
    // No? Reset reader and Advance(position)
}
```

In theory the common case would be backtracking in the current span. Another option:

``` C#
public void Backtrack(long count)
{
    // Check for > 0
    // Is count in current segment?
    // Yes? Calculate CurrentSpanPosition and set & set Consumed
    // No? Reset reader and Advance(position)
}
```

@KrzysztofCwalina, thoughts?