@ahsonkhan 2018-10-04
> Is it possible to backtrack across segment boundaries? i.e. don't we lose past segments?

Right. Even though we have the original ReadOnlySequence currently, we would have to move forward from the start of the sequence to the back-track point. That's because there is no way to back-track the underlying ReadOnlySequence (maybe something we should consider adding).

And if we don't keep the original ReadOnlySequence, then there is no way to backtrack across the segment boundaries.