@JeremyKuhne 2018-10-04
> Is it possible to backtrack across segment boundaries? i.e. don't we lose past segments?

You don't, otherwise we'd be busted all over (you wouldn't be able to save a copy and continue). If you go to Pipe and grab another sequence you need to start all over in that case.