@JeremyKuhne 2018-10-04
> there is no way to back-track the underlying ReadOnlySequence (maybe something we should consider adding).

You'd have to build a reverse linked list on the fly, which doesn't seem plausible or particularly useful. We could keep the prior SequencePosition if we thought that it would be frequently useful- but that would bloat us further.