@ahsonkhan 2018-10-05
**We went from 1.32x to 1.96x down to 1.58x. So were able to get back most of the regression :)**

Most of the 10% cost is to support the re-entrant reader which was expected.

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17754
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009460
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT


```

**Master:**

|                Method |  TestCase | segmentSize |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|---------------------- |---------- |------------ |-----------:|---------:|---------:|-------:|----------:|
|  **MultiSegmentSequence** | **Json400KB** |        **4000** | **1,275.6 us** | **8.582 us** | **7.166 us** | **1.9531** |   **14072 B** |
| **SingleSegmentSequence** | **Json400KB** |           **?** |   **963.2 us** | **5.585 us** | **5.224 us** |      **-** |       **0 B** |

**Before:**

|                Method |  TestCase | segmentSize |       Mean |     Error |    StdDev | Allocated |
|---------------------- |---------- |------------ |-----------:|----------:|----------:|----------:|
|  **MultiSegmentSequence** | **Json400KB** |        **4000** | **1,900.5 us** | **17.172 us** | **16.063 us** |   **14072 B** |
| **SingleSegmentSequence** | **Json400KB** |           **?** |   **969.2 us** |  **9.211 us** |  **7.692 us** |       **0 B** |


**Now:**

|                Method |  TestCase | segmentSize |       Mean |     Error |   StdDev |  Gen 0 | Allocated |
|---------------------- |---------- |------------ |-----------:|----------:|---------:|-------:|----------:|
|  **MultiSegmentSequence** | **Json400KB** |        **4000** | **1,391.3 us** | **10.106 us** | **9.453 us** | **1.9531** |   **14072 B** |
| **SingleSegmentSequence** | **Json400KB** |           **?** |   **879.4 us** |  **3.936 us** | **3.287 us** |      **-** |       **0 B** |